telegram_bot/requirements.txt
ipdb==0.13.6
ipython==7.21.0
python-telegram-bot==13.3
requests==2.25.1
python-dotenv==0.15.0